refactor: Project structure.
Split the sign tool into several files for expandability.
